Crude logic implementation
Templates will have a section which contains logic that can be evaluated in order to make dynamic changes to the output.  This is represented by what are effectively code blocks which have backing expression classes.  The overall goal of this is to allow for easy plugin creation in order to expose APIs from other services which can be invoked through these logic steps.