Add type alias for LogicContext
LogicContext is effectively an empty dict which is created when a PosterTemplate is constructed.  It is meant to be a workspace for all of the logic steps to play with and will be prepopulated with the passed in args.  In this way, the logic steps have a way to store and retrieve data.