Fix unintentional expressions mutation
Having Plugin.expression as a static method meant __class__ always returned Plugin rather than the invoking class; instead, Plugin.expression is now a class method.  Also, having a list as a class member results in the same list being shared among all subclasses, so Plugin.expression now builds a new one behind the scenes which can be accessed by the Plugin.expressions property.  Lastly, registering an expression requires the class to be the Plugin it belongs to, not just Plugin.