Fix plugin contexts
Originally only a specific plugin's context would get passed to each expression.  However, if that expression tried to evaluate another expression that was from a different plugin, there was no way to pass that other plugin's context.  Now there's a ContextProvider class that handles this; it gets passed to each Expression object in evaluate() and has a method for getting the current Expression's context.  That way the ContextProvider can be passed to nested Expressions too.