Leverage pyvips for Element rendering
Rather than using PIL, this project will work to leverage pyvips for rendering posters since image manipulation plays a big part in the scope of this project.  With a previous attempt at this project, a PIL-based renderer took about ~10 seconds to composite a 6000x9000px poster from start to finish.  That's a long time!  Hopefully with the power of pyvips, that time can be drastically reduced to make this a realistically usable cli utility.