Add REQUIRED constant for Expression signatures
The model for an Expression originally defined evaluate() as Callable[..., T] where T is the evaluated type.  This leaves a lot of room for plugin implementation error, namely due to lack of static typechecking when writing an Expression's evaluate method.  The method is now defined with a concrete signature, but Pylance doesn't seem to allow for **kwargs to be forwarded to subclasses correctly (and I can see why?  e.g. casting Child to Parent and invoking the method will be missing parameters).  As a workaround, a constant REQUIRED (an Any-casted singleton) is defined in the API and can be dropped in as the default value for any required args when evaluating an Expression.  This can be omitted from context.  This achieves the following:
- Subclassing Expression is less of a pain with evaluate() being recognized as a valid classmember
- Overriding Expression.evaluate can be autofilled by Pylance
- Type signature of overridden evaluate can be statically checked for correctness